fix every thing for sql and oracle in Ecm

+++ Data/Data/ECM/ArtSystemPerformance.cs
@@ -23,9 +23,9 @@
         public string? LockedBy { get; set; }
         public DateTime? EcmLastStatusDate { get; set; }
         public string? HitsCount { get; set; }
-        public decimal? DurationsInSeconds { get; set; }
-        public decimal? DurationsInMinutes { get; set; }
-        public decimal? DurationsInHours { get; set; }
-        public decimal? DurationsInDays { get; set; }
+        public int? DurationsInSeconds { get; set; }
+        public int? DurationsInMinutes { get; set; }
+        public int? DurationsInHours { get; set; }
+        public int? DurationsInDays { get; set; }
     }
 }
